[ruby] add counters in a test.
